+++ src/reply.rs
@@ -12,7 +12,7 @@ pub enum Reply {
         description: String,
         payload: String,
         currency: String,
-        amount: u64,
+        amount: f64,
     },
 }
 #[derive(Debug)]
